# packages/api-client-go/api/openapi.yaml
openapi: 3.0.0
info:
  contact: {}
  description: Snapflow AI platform API Docs
  title: Snapflow
  version: "1.0"
servers:
- url: http://localhost:3000
paths:
...
  /api-keys:
    get:
      operationId: listApiKeys
      parameters:
      - description: Use with JWT to specify the organization ID
        explode: false
        in: header
        name: X-Snapflow-Organization-ID
        required: false
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ApiKeyList'
                type: array
          description: API keys retrieved successfully.
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: List API keys
      tags:
      - api-keys
    post:
      operationId: createApiKey
      parameters:
      - description: Use with JWT to specify the organization ID
        explode: false
        in: header
        name: X-Snapflow-Organization-ID
        required: false
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApiKey'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyResponse'
          description: API key created successfully.
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Create API key
      tags:
      - api-keys
  /api-keys/current:
    get:
      operationId: getCurrentApiKey
      parameters:
      - description: Use with JWT to specify the organization ID
        explode: false
        in: header
        name: X-Snapflow-Organization-ID
        required: false
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyList'
          description: API key retrieved successfully.
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Get current API key's details
      tags:
      - api-keys
  /api-keys/{name}:
    delete:
      operationId: deleteApiKey
      parameters:
      - description: Use with JWT to specify the organization ID
        explode: false
        in: header
        name: X-Snapflow-Organization-ID
        required: false
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "204":
          description: API key deleted successfully.
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Delete API key by name
      tags:
      - api-keys
    get:
      operationId: getApiKey
      parameters:
      - description: Use with JWT to specify the organization ID
        explode: false
        in: header
        name: X-Snapflow-Organization-ID
        required: false
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: name
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiKeyList'
          description: API key retrieved successfully.
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Get API key by name
      tags:
      - api-keys
...
  /organizations/{organizationId}/billing/checkout:
    get:
      operationId: createCheckoutLink
      parameters:
      - explode: false
        in: path
        name: organizationId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationCheckout'
        required: true
      responses:
        "200":
          description: ""
      security:
      - bearer: []
      - oauth2:
        - openid
        - profile
        - email
      summary: Get payment checkout link
      tags:
      - organizations
...
components:
  schemas:
...
    CreateApiKey:
      example:
        permissions:
        - write:registries
        - write:registries
        name: Test API Key
        expiresAt: 2000-01-23T04:56:07.000+00:00
      properties:
        name:
          description: The name of the API key
          example: Test API Key
          type: string
        permissions:
          description: The list of organization resource permissions assigned to the
            API key
          items:
            enum:
            - write:registries
            - delete:registries
            - write:images
            - delete:images
            - write:sandboxes
            - delete:sandboxes
            - read:buckets
            - write:buckets
            - delete:buckets
            type: string
          type: array
        expiresAt:
          description: When the API key is set to expire
          format: date-time
          nullable: true
          type: string
      required:
      - name
      - permissions
      type: object
    ApiKeyResponse:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        permissions:
        - write:registries
        - write:registries
        name: Test API Key
        value: snapflow_1234567890abcdef
        expiresAt: 2000-01-23T04:56:07.000+00:00
      properties:
        name:
          description: The name of the API key
          example: Test API Key
          type: string
        value:
          description: The value of the API key
          example: snapflow_1234567890abcdef
          type: string
        permissions:
          description: List of permissions given to the API key
          items:
            enum:
            - write:registries
            - delete:registries
            - write:images
            - delete:images
            - write:sandboxes
            - delete:sandboxes
            - read:buckets
            - write:buckets
            - delete:buckets
            type: string
          type: array
        createdAt:
          description: When the API key was made
          format: date-time
          type: string
        expiresAt:
          description: When the API key is set to expire
          format: date-time
          nullable: true
          type: string
      required:
      - createdAt
      - expiresAt
      - name
      - permissions
      - value
      type: object
    ApiKeyList:
      example:
        createdAt: 2000-01-23T04:56:07.000+00:00
        lastUsedAt: 2000-01-23T04:56:07.000+00:00
        permissions:
        - write:registries
        - write:registries
        name: Test API Key
        value: snapflow_************************
        expiresAt: 2000-01-23T04:56:07.000+00:00
      properties:
        name:
          description: The name of the API key
          example: Test API Key
          type: string
        value:
          description: The hidden API key value
          example: snapflow_************************
          type: string
        permissions:
          description: List of permissions given to the API key
          items:
            enum:
            - write:registries
            - delete:registries
            - write:images
            - delete:images
            - write:sandboxes
            - delete:sandboxes
            - read:buckets
            - write:buckets
            - delete:buckets
            type: string
          type: array
        createdAt:
          description: When the API key was made
          format: date-time
          type: string
        expiresAt:
          description: When the API key is set to expire
          format: date-time
          nullable: true
          type: string
        lastUsedAt:
          description: When the latest request was made using the API key
          format: date-time
          nullable: true
          type: string
      required:
      - createdAt
      - expiresAt
      - lastUsedAt
      - name
      - permissions
      - value
      type: object
...
    CreateOrganizationCheckout:
      example:
        productId: d13c586c-ce41-4673-8917-0043c0a22867
      properties:
        productId:
          description: The product ID of the subscription that is being purchased
          example: d13c586c-ce41-4673-8917-0043c0a22867
          type: string
      required:
      - productId
      type: object
...
  securitySchemes:
    bearer:
      bearerFormat: JWT
      description: API Key access
      scheme: bearer
      type: http